issue #123: Make Facts#get method cast friendly

+++ easy-rules-tutorials/src/main/java/org/jeasy/rules/tutorials/airco/AirConditioningRule.java
@@ -17,7 +17,7 @@ public class AirConditioningRule {
     @Action
     public void coolAir(Facts facts) {
         System.out.println("It is hot! cooling air..");
-        Integer temperature = (Integer) facts.get("temperature");
+        Integer temperature = facts.get("temperature");
         facts.put("temperature", temperature - 1);
     }
